:art: refactor: Split icon generator into 3 separate files
